init: Add --merge-base to explicitly specify a merge base
Using an automatic merge base computed with git-merge-base together with
log --ancestry-path may generate unexpectedly few merge points.

This commit adds an option that solves half of the problem:  It lets one
explicitly specify a commit to be used as the merge based.

The second part of the problem is --ancestry-path.  I observed that
a larger series of first-parent commits that were not all in the
ancestry path (around 100 commits) was easier to reason about than the
fewer ancestry-path commits (around 10 commits).  I did that merge by
manually disabling the --ancestry-path filter for tip2.